requirements.txt
spacy==3.1.3
pandas==1.3.1
azure-core==1.16.0
ipyfilechooser==0.6.0
azureml-sdk==1.33.0
openpyxl==3.0.7